backend/ToDoApp/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
PyJWT==1.6.3
pytz==2018.4
